Removed two references to the head of the times seq which caused a memory leak
I use your nice library and observed a memory leak in my programs. 
A huge number of org.joda.time.DateTime objects was found in the old gen space of my program.
After a deeper inspection of the code I have seen that there are two references to the head element of the "times"-seq.

The first one is implicit in the destructuring of [next-time & more-times] in the go-loop, see https://groups.google.com/forum/#!topic/clojure/YNcqmAwWkqA .
I fixed it by removing the destructuring and use instead first and rest in the body.

The second one is the function itself. The reify returns an object that holds (amongst others) a reference to the function arguments, in particular to times.
I fixed it by converting the functions to macros.

Unfortunately you still have to be careful in using the lib.
For instance
(let [times (periodic/periodic-seq (time/now) (time/millis 1000))]
  (chime-ch times))
still produces a leak, while
(chime-ch (periodic/periodic-seq (time/now) (time/millis 1000)))
works fine.

It would be great if you would accept my proposal and adapt the use-cases in the readme.